storing marker coordinates from marker in marker_coordiantes and placing popup on said coordiantes.
