+++ members/static/js/maps.js
@@ -1,4 +1,6 @@
 import { RESTAURANTS, TOURIST_ATTRACTIONS, HOTELS } from "./placeTypes.js";
+let map, popup, Popup;
+let marker_coordinates = [];
 
 function handleSave(itenarary_item) {
   itenarary_saves.push(itenarary_item);
@@ -9,13 +11,13 @@ function handleSave(itenarary_item) {
     p.innerHTML += `<strong><u>${key}</u></strong>: ${itenarary_item[key]} <br>`;
   }
   div.appendChild(p);
-  // div.appendChild(new_line);
   div.className = "itenarary_item";
   list.appendChild(div);
 }
 
 function getLocationInfo(location) {
   let locationTypes = [];
+  // console.log(locationInformation);
   for (let type of location.types) {
     if (
       type === "restaurant" ||
@@ -29,10 +31,10 @@ function getLocationInfo(location) {
     locationTypes.push(location.types[0]);
   }
 
+  console.log(location);
   return {
-    address: location.vicinity,
+    address: location.vicinity ? location.vicinity : location.formatted_address,
     name: location.name,
-    // types: location.types.splice(0, 2),
     total_num_ratings: location.user_ratings_total,
     rating: location.rating,
     type: locationTypes,
@@ -42,8 +44,9 @@ function getLocationInfo(location) {
 console.log(RESTAURANTS);
 let itenarary_saves = [];
 let places;
-let infoWindow;
-let autocomplete;
+//>>>>
+// let infoWindow;
+// let autocomplete;
 let markers = [];
 const MARKER_PATH =
   "https://developers.google.com/maps/documentation/javascript/images/marker_green";
@@ -55,7 +58,7 @@ function displayDropDown() {
 window.addEventListener("click", function (event) {
   if (!event.target.matches("#select-button")) {
     let dropdowns = document.getElementsByClassName("dropdown-buttons");
-    for (dropdown of dropdowns) {
+    for (let dropdown of dropdowns) {
       let openDropDown = dropdown;
       if (openDropDown.classList.contains("display")) {
         openDropDown.classList.remove("display");
@@ -74,24 +77,84 @@ function initMap() {
     zoom: 14,
     center: startLatLng,
   };
+  //credit: https://developers.google.com/maps/documentation/javascript/examples/overlay-popup
+  /**
+   * A customized popup on the map.
+   */
+  class Popup extends google.maps.OverlayView {
+    position;
+    containerDiv;
+    constructor(position, content) {
+      super();
+      this.position = position;
+      content.classList.add("popup-bubble");
+
+      // This zero-height div is positioned at the bottom of the bubble.
+      const bubbleAnchor = document.createElement("div");
+
+      bubbleAnchor.classList.add("popup-bubble-anchor");
+      bubbleAnchor.appendChild(content);
+
+      // This zero-height div is positioned at the bottom of the tip.
+      this.containerDiv = document.createElement("div");
+      this.containerDiv.classList.add("popup-container");
+      this.containerDiv.appendChild(bubbleAnchor);
+
+      // Optionally stop clicks, etc., from bubbling up to the map.
+      Popup.preventMapHitsAndGesturesFrom(this.containerDiv);
+    }
+    /** Called when the popup is added to the map. */
+    onAdd() {
+      this.getPanes().floatPane.appendChild(this.containerDiv);
+    }
+    /** Called when the popup is removed from the map. */
+    onRemove() {
+      if (this.containerDiv.parentElement) {
+        this.containerDiv.parentElement.removeChild(this.containerDiv);
+      }
+    }
+    /** Called each frame when the popup needs to draw itself. */
+    draw() {
+      const divPosition = this.getProjection().fromLatLngToDivPixel(
+        this.position
+      );
+      // Hide the popup when it is far out of view.
+      const display =
+        Math.abs(divPosition.x) < 4000 && Math.abs(divPosition.y) < 4000
+          ? "block"
+          : "none";
+
+      if (display === "block") {
+        this.containerDiv.style.left = divPosition.x + "px";
+        this.containerDiv.style.top = divPosition.y + "px";
+      }
+
+      if (this.containerDiv.style.display !== display) {
+        this.containerDiv.style.display = display;
+      }
+    }
+  }
+
+  // let popupContent = document.createElement("div");
+  // let searchSaveButton;
+  // popupcContent.innerHTML = "";
+  // popup = new Popup(
+  //   new google.maps.LatLng(-33.866, 151.196),
+  //   document.getElementById("content")
+  // );
+  // popup.setMap(map);
+
   const map = new google.maps.Map(document.getElementById("map"), options);
   places = new google.maps.places.PlacesService(map);
-  // console.log(">>> places", places.getDetails());
   let location_info = [];
-  let index = 0;
-
-  google.maps.event.addListener(map, "click", function (event) {
-    addMarker({ coords: event.latLng });
-    location_info.push({
-      // index: [{ lat: event.latLng.lat() }, { lng: event.latLng.lng() }],
-    });
-    index += 1;
-    console.log(location_info);
-  });
 
   //search bar
   let input = document.getElementById("search");
+  let clear = document.getElementById("clear-button");
   let searchBox = new google.maps.places.SearchBox(input);
+  clear.addEventListener("click", () => {
+    input.value = "";
+  });
   //makes sure that the search is only limited to the bounds of the map box
   map.addListener("bounds_changed", function () {
     searchBox.setBounds(map.getBounds());
@@ -110,22 +173,21 @@ function initMap() {
     });
     //re-init to an empty array
     markers = [];
-
     let bounds = new google.maps.LatLngBounds();
     places.forEach(function (place) {
       if (!place.geometry) {
         return;
       }
-      markers.push(
-        new google.maps.Marker({
-          map: map,
-          title: place.name,
-          position: place.geometry.location,
-          // content: `lat: ${place.geometry.location.lat()} <br> lng: ${place.geometry.location.lng()}`,
-          content: ` ${place.name} <br>  ${place.vicinity} 
-          <br> <button onclick="handleSave(getLocationInfo(place))">Save</button>`,
-        })
-      );
+      console.log(">>> jy place", place);
+      let marker = new google.maps.Marker({
+        map: map,
+        title: place.name,
+        position: place.geometry.location,
+        placeResult: place,
+      });
+      marker.content = createPopupContent(marker, marker.placeResult);
+
+      markers.push(marker);
 
       console.log("coordinates", place.geometry.location.lat());
       if (place.geometry.viewport) {
@@ -135,40 +197,39 @@ function initMap() {
       }
     });
     //>>>>>
+
+    // markers[i].placeResult = results[i];
+    // google.maps.event.addListener(markers[i], "click", showInfoWindow);
+    // google.maps.event.addListener(marker, "click", () =>
+    //   displayInfoWindow(marker, place)
+    // );
+
+    // let marker_coordinates = [];
     markers.forEach((marker) => {
-      let infoWindow = new google.maps.InfoWindow({
-        //  content: "lat: " + lat + "<br/>lng: " + lng,
-        content: marker.content,
-        // content: document.getElementById(filterTextbox),
+      google.maps.event.addListener(marker, "click", function (event) {
+        marker_coordinates.push(event.latLng.lat());
+        marker_coordinates.push(event.latLng.lng());
+        console.log(marker_coordinates);
       });
       marker.addListener("click", function () {
-        infoWindow.open(map, marker);
+        // displayInfoWindow(marker, place);
+        // console.log(">>>>inside marker", marker, "placeREsult", marker.content);
+        // let popupInfoWindow = new Popup{
+        //   marker.position,
+        // };
+        displayPopup(marker, marker.placeResult);
       });
+      // marker.addListener("click", function () {
+      //   infoWindow.open(map, marker);
+      // });
       location_info.push({
         index: [{ lat: marker.lat }, { lng: marker.lng }],
       });
-      index += 1;
-      console.log(location_info);
+      // index += 1;
+      console.log("marker bounds test", location_info);
     });
     map.fitBounds(bounds);
   });
-
-  function addMarker(props) {
-    console.log("this is props", props);
-    let string_coords = String(props.coords);
-    let coordinates = string_coords.split(", ");
-    let lat = coordinates[0].replace("(", "");
-    let lng = coordinates[1].replace(")", "");
-    console.log("lng is " + lng + " and lat is " + lat);
-    var marker = new google.maps.Marker({
-      position: props.coords,
-      map,
-    });
-
-    marker.addListener("click", function () {
-      infoWindow.open(map, marker);
-    });
-  }
   //credit: https://developers.google.com/maps/documentation/javascript/examples/places-autocomplete-hotelsearch
   function dropMarker(i) {
     return function () {
@@ -186,8 +247,7 @@ function initMap() {
     markers = [];
   }
 
-  function search(type) {
-    // console.log("in search");
+  function filter(type) {
     const search = {
       bounds: map.getBounds(),
       types: [type],
@@ -209,37 +269,75 @@ function initMap() {
           });
           // If the user clicks a hotel marker, show the details of that hotel
           // in an info window.
-          // @ts-ignore TODO refactor to avoid storing on marker
           markers[i].placeResult = results[i];
-          // google.maps.event.addListener(markers[i], "click", showInfoWindow);
-          google.maps.event.addListener(markers[i], "click", displayInfoWindow);
+          // google.maps.event.addListener(markers[i], "click", displayInfoWindow);
+          google.maps.event.addListener(markers[i], "click", function (event) {
+            marker_coordinates.push(event.latLng.lat());
+            marker_coordinates.push(event.latLng.lng());
+            console.log("filter marker coordinates: ", marker_coordinates);
+          });
+          google.maps.event.addListener(markers[i], "click", () =>
+            displayPopup(markers[i], results[i])
+          );
+
           setTimeout(dropMarker(i), i * 100);
           addResult(results[i], i);
         }
-        // console.log(">> saves", itenarary_saves);
       }
     });
+  }
+
+  function createPopupContent(marker, place) {
+    let container = document.createElement("div");
+    let button = document.createElement("button");
+    button.innerText = "Save";
+    button.className = "filterSaveButton";
+    button.addEventListener("click", () => {
+      let info = getLocationInfo(place);
+      handleSave(info);
+    });
+
+    let { name, rating, user_ratings_total, vicinity, formatted_address } =
+      marker.placeResult;
+
+    container.className = "popup-content";
+    container.innerHTML = `${name} <br> average rating: ${rating} <br> total number of user ratings:  ${user_ratings_total} <br> ${
+      vicinity ? vicinity : formatted_address
+    } <br> `;
+    container.appendChild(button);
+    return container;
+  }
+
+  function displayPopup(marker, place) {
+    let content = createPopupContent(marker, place);
+    console.log("marker position", marker.position.lat);
+    let popup = new Popup(marker.position, content);
+    map.addListener("click", function () {
+      popup.setMap(null);
+    });
+    popup.setMap(map);
   }
 
   let restaurantButton = document.getElementById("restaurants");
   let hotelsButton = document.getElementById("hotels");
   let touristAttractionsButton = document.getElementById("tourist_attractions");
-  restaurantButton.addEventListener("click", () => search(RESTAURANTS));
-  hotelsButton.addEventListener("click", () => search(HOTELS)); //
+  restaurantButton.addEventListener("click", () => filter(RESTAURANTS));
+
+  hotelsButton.addEventListener("click", () => filter(HOTELS)); //>>>>
   touristAttractionsButton.addEventListener("click", () =>
-    search(TOURIST_ATTRACTIONS)
+    filter(TOURIST_ATTRACTIONS)
   );
+  //>>>
 }
 function displayInfoWindow() {
   let marker = this;
-  // console.log("marker", marker);
   let infoWindow = new google.maps.InfoWindow({
-    content: ` ${marker.placeResult.name} <br> average rating: ${marker.placeResult.rating} <br> total number of user ratings:  ${marker.placeResult.user_ratings_total} <br> ${marker.placeResult.vicinity}`,
+    content: ` ${marker.placeResult.name} <br> average rating: ${marker.placeResult.rating} <br> total number of user ratings:  ${marker.placeResult.user_ratings_total} <br> ${marker.placeResult.vicinity} `,
   });
   infoWindow.open(map, marker);
 }
 
-//credit: https://developers.google.com/maps/documentation/javascript/examples/places-autocomplete-hotelsearch
+// credit: https://developers.google.com/maps/documentation/javascript/examples/places-autocomplete-hotelsearch
 function clearResults() {
   const results = document.getElementById("results");
 
@@ -270,8 +368,6 @@ function addResult(result, i) {
 
   const name = document.createTextNode(result.name);
   let itenarary_item = getLocationInfo(result);
-  // console.log(itenarary_item);
-  iconTd.appendChild(icon);
   nameTd.appendChild(name);
   tr.appendChild(iconTd);
   tr.appendChild(nameTd);
