+++ project/static/loginstylesheet.css
@@ -1,0 +1,419 @@
+/* Global Styles */
+body {
+  background-color: #f0e6d6;
+  width: 100%;
+  height: 100%;
+  font-family: 'Source Sans Pro', sans-serif;
+  margin: 0;
+  display: flex;
+  flex-direction: column;
+  margin-bottom: 0%;
+}
+li, li::marker {
+  list-style-type: none;
+}
+
+/* Header Styles */
+header {
+  background-color: #E7C176;
+  display: flex;
+  justify-content: space-between;
+  padding: 5px;
+  box-shadow: 0px 2px 5px rgba(0, 0, 0, 0.2);
+}
+
+nav {
+  height: 75px;
+}
+
+nav ul {
+  list-style-type: none;
+  padding: 0;
+  margin: 0;
+  overflow: hidden;
+}
+
+nav li {
+  float: right;
+}
+
+nav li a {
+  display: block;
+  color: rgb(10, 8, 8);
+  text-align: center;
+  padding: 14px 16px;
+  text-decoration: none;
+  font-size: 20px;
+  font-weight: bolder;
+}
+
+nav li a:hover {
+  background-color: #e8ba5c;
+  border-radius: 10px;
+  transition: background-color 0.3s ease;
+}
+
+/* Footer Styles */
+#feet {
+  background-color: #E7C176;
+  color: #fff;
+  padding: 10px;
+  margin-bottom: 0%;
+}
+
+.container-footer {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  font-size: 18px;
+}
+
+.footer-links {
+  display: flex;
+}
+
+.footer-links a {
+  display: block;
+  color: rgb(4, 2, 2);
+  text-align: center;
+  padding: 14px 16px;
+  text-decoration: none;
+  font-weight: bold;
+  font-size: larger;
+}
+
+.footer-links a:hover {
+  background-color: #e8ba5c;
+  border-radius: 10px;
+  transition: background-color 0.3s ease;
+}
+
+.footer-logo {
+  display: flex;
+  align-items: center;
+  margin-right: 2%;
+}
+
+.footer-logo img {
+  margin-right: 0px;
+}
+
+/* Login Styles */
+.login-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  /* margin-top: 4%;
+  margin-bottom: 4%; */
+  margin: 5%;
+}
+
+.login-box {
+  background-color: #fff;
+  border-radius: 20px;
+  box-shadow: 0px 2px 5px rgba(0, 0, 0, 0.2);
+  padding: 30px;
+  width: 300px;
+}
+
+.lock-icon {
+  display: flex;
+  justify-content: center;
+  margin-bottom: 30px;
+}
+
+.lock-icon img {
+  width: 100px;
+  height: 100px;
+}
+
+h2 {
+  text-align: center;
+  font-size: 28px;
+  margin-bottom: 20px;
+}
+
+form {
+  text-align: center;
+  display: flex;
+  flex-direction: column;
+}
+
+.input-container {
+  display: flex;
+  flex-direction: column;
+  margin-bottom: 20px;
+}
+
+label {
+  margin-bottom: 5px;
+  font-size: 18px;
+  font-weight: bold;
+}
+
+input {
+  border-radius: 5px;
+  height: 40px;
+  padding: 5px;
+  font-size:larger;
+}
+
+button{
+  background-color: #e8ba5c;
+color: #fff;
+border: none;
+padding: 8px 16px;
+border-radius: 10px;
+cursor: pointer;
+margin-left: 8px;
+  font-size: larger;
+}
+ /* Signup Styles */
+ .signup-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-top: 3%;
+  margin-bottom: 3%;
+}
+
+.signup-box {
+  background-color: #fff;
+  border-radius: 20px;
+  box-shadow: 0px 2px 5px rgba(0, 0, 0, 0.2);
+  padding: 30px;
+  width: 350px;
+}
+
+.lock-icon {
+  display: flex;
+  justify-content: center;
+  margin-bottom: 30px;
+}
+
+.lock-icon img {
+  width: 100px;
+  height: 100px;
+}
+
+h2 {
+  text-align: center;
+  font-size: 28px;
+  margin-bottom: 20px;
+}
+
+form {
+  text-align: center;
+  display: flex;
+  flex-direction: column;
+}
+
+.input-container {
+  display: flex;
+  flex-direction: column;
+  margin-bottom: 5px;
+}
+
+label {
+  margin-bottom: 5px;
+  font-size: 16px;
+  font-weight: bold;
+}
+
+input {
+  border-radius: 5px;
+  height: 20px;
+  padding: 5px;
+  font-size:larger;
+}
+
+button{
+  background-color: #e8ba5c;
+color: #fff;
+border: none;
+padding: 8px 16px;
+border-radius: 10px;
+cursor: pointer;
+margin-left: 8px;
+  font-size: larger;
+}
+.helptext {
+  visibility: hidden;
+}
+form span{
+  display: none;
+}
+form ul{
+  display: none;
+}
+form p:nth-child(4) {
+  height: 0;
+  width: 0;
+  padding: 0;
+  margin: 0;
+  border: none;
+}
+p:empty {
+  display: none;
+}
+form p{
+  display: flex;
+  flex-direction: column;
+}
+
+/* Main Page Styles */
+.dropdown-buttons {
+	display: none;
+	position: absolute;
+	z-index: 1;
+}
+
+.display {
+	display: block;
+}
+#map {
+	height: calc(100vh - 200px); /* subtract the height of navigation bar and search bar from viewport height */
+	width: 100%;
+	grid-row: 1;
+	border-radius: 10px;
+	border-color: #E7C176;
+}
+  
+#search {
+	border-radius: 10px;
+	width: 100%;
+	/* height: 40px; */
+	padding: 8px 16px;
+}
+.map{
+	border-radius: 10px;
+	border-color: #E7C176;
+}
+.container {
+    margin-top: 5px;
+    margin-bottom: 100px;
+    display: grid;
+    grid-template-columns: 2fr 1fr;
+    grid-row: 1 / -1;;
+    height: calc(100vh - 200px); /* subtract the height of navigation bar and search bar from viewport height */
+    padding: 0px 25px ;
+}
+.textbox {
+      border-radius: 10px;
+      margin-left: 20px;
+      height: 100%;
+      grid-column: 2;
+      grid-row: 1;
+      background-color: #E7C176;
+      overflow-y: auto; /* add a scrollbar if the content overflows */
+}
+h2{
+	padding-left: 10px;
+}
+.container-heading{
+	padding-top: 5px;
+	display: flex;
+	justify-content: space-between;
+	padding-left: 1%;
+	padding-right: 1%;
+}
+.search-container {
+	margin: 3px;
+	display: flex;
+	align-items: center;
+	border: 2px solid #ccc;
+	border-radius: 20px;
+	padding: 7px;
+}
+  
+.search-container input[type="text"] {
+	border: none;
+	outline: none;
+	font-size: 16px;
+	padding: 5px;
+	flex: 1;
+}
+  
+.search-container button[type="submit"] {
+	background-color: #E7C176;
+	color: #fff;
+	border: none;
+	border-radius: 20px;
+	padding: 10px;
+	cursor: pointer;
+	margin-left: 10px;
+}
+  
+.search-container button[type="submit"]:hover {
+	background-color: #9dbb9f;
+}
+
+/* #select-button {
+	background-color: #e8ba5c;
+	color: #fff;
+	border: none;
+	padding: 8px 16px;
+	border-radius: 4px;
+	cursor: pointer;
+	margin-left: 8px;
+} */
+  
+#clear-button {
+	background-color: #dc3545;
+	color: #fff;
+	border: none;
+	padding: 8px 16px;
+	border-radius: 4px;
+	cursor: pointer;
+	margin-left: 8px;
+}
+
+.filterContainer {
+  position: relative;
+  margin-left: 10px;
+}
+
+#select-button {
+  background-color: #E7C176;
+	color: #040303;
+  font-weight: bold;
+	border: none;
+	padding: 8px 16px;
+	border-radius: 4px;
+	cursor: pointer;
+	margin-left: 4px;
+}
+
+.dropdown-buttons {
+  display: none;
+  position: absolute;
+  top: 100%;
+  left: 0;
+  width: 200px;
+  background-color: #fce0a8;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+  padding: 10px;
+  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
+  margin-top: 2% ;
+}
+
+.dropdown-buttons button {
+  display: block;
+  width: 100%;
+  padding: 8px;
+  background-color: #ad5e04;
+  margin: 2%;
+  border: none;
+  text-align: left;
+  cursor: pointer;
+  transition: background-color 0.3s ease;
+}
+
+.dropdown-buttons button:hover {
+  background-color: #363333;
+}
+
+.filterContainer:hover .dropdown-buttons {
+  display: block;
+}
